Added icon & button styling for elements
