+++ src/core/components/ElementButton/index.js
@@ -1,0 +1,3 @@
+import ElementButton from './ElementButton';
+
+export default ElementButton;
