+++ src/core/components/ElementIcon/index.js
@@ -1,0 +1,3 @@
+import ElementIcon from './ElementIcon';
+
+export default ElementIcon;
