--- src/gear-preview/components/Element/index.js
@@ -1,3 +1,0 @@
-import Element from './Element';
-
-export default Element;
